TaskUIController_xxx_Testをエンクロージング・クラスとして取り込むため、TaskUIControllerTestに取り込みました。
